Draft: split PointArray from Draft.py
