+++ src/Mod/Draft/Draft.py
@@ -307,6 +307,10 @@ if FreeCAD.GuiUp:
 # arrays
 from draftmake.make_patharray import make_path_array, makePathArray
 from draftobjects.patharray import PathArray, _PathArray
+
+from draftmake.make_pointarray import make_point_array, makePointArray
+from draftobjects.pointarray import PointArray, _PointArray
+
 if FreeCAD.GuiUp:
     from draftviewproviders.view_array import ViewProviderDraftArray
     from draftviewproviders.view_array import _ViewProviderDraftArray
@@ -475,21 +479,6 @@ def makeArray(baseobject,arg1,arg2,arg3,arg4=None,arg5=None,arg6=None,name="Arra
             if len(obj.Base.ViewObject.DiffuseColor) > 1:
                 obj.ViewObject.Proxy.resetColors(obj.ViewObject)
         baseobject.ViewObject.hide()
-        select(obj)
-    return obj
-
-def makePointArray(base, ptlst):
-    """makePointArray(base,pointlist):"""
-    obj = FreeCAD.ActiveDocument.addObject("Part::FeaturePython","PointArray")
-    _PointArray(obj, base, ptlst)
-    obj.Base = base
-    obj.PointList = ptlst
-    if gui:
-        _ViewProviderDraftArray(obj.ViewObject)
-        base.ViewObject.hide()
-        formatObject(obj,obj.Base)
-        if len(obj.Base.ViewObject.DiffuseColor) > 1:
-            obj.ViewObject.Proxy.resetColors(obj.ViewObject)
         select(obj)
     return obj
 
@@ -866,57 +855,6 @@ class _Array(_DraftLink):
                     npl.translate(FreeCAD.Vector(axisvector).multiply(i+1))
             base.append(npl)
         return base
-
-class _PointArray(_DraftObject):
-    """The Draft Point Array object"""
-    def __init__(self, obj, bobj, ptlst):
-        _DraftObject.__init__(self,obj,"PointArray")
-        obj.addProperty("App::PropertyLink","Base","Draft",QT_TRANSLATE_NOOP("App::Property","Base")).Base = bobj
-        obj.addProperty("App::PropertyLink","PointList","Draft",QT_TRANSLATE_NOOP("App::Property","PointList")).PointList = ptlst
-        obj.addProperty("App::PropertyInteger","Count","Draft",QT_TRANSLATE_NOOP("App::Property","Count")).Count = 0
-        obj.setEditorMode("Count", 1)
-
-    def execute(self, obj):
-        import Part
-        from FreeCAD import Base, Vector
-        pls = []
-        opl = obj.PointList
-        while getType(opl) == 'Clone':
-            opl = opl.Objects[0]
-        if hasattr(opl, 'Geometry'):
-            place = opl.Placement
-            for pts in opl.Geometry:
-                if hasattr(pts, 'X') and hasattr(pts, 'Y') and hasattr(pts, 'Z'):
-                    pn = pts.copy()
-                    pn.translate(place.Base)
-                    pn.rotate(place)
-                    pls.append(pn)
-        elif hasattr(opl, 'Links'):
-            pls = opl.Links
-        elif hasattr(opl, 'Components'):
-            pls = opl.Components
-
-        base = []
-        i = 0
-        if hasattr(obj.Base, 'Shape'):
-            for pts in pls:
-                #print pts # inspect the objects
-                if hasattr(pts, 'X') and hasattr(pts, 'Y') and hasattr(pts, 'Z'):
-                    nshape = obj.Base.Shape.copy()
-                    if hasattr(pts, 'Placement'):
-                        place = pts.Placement
-                        nshape.translate(place.Base)
-                        nshape.rotate(place.Base, place.Rotation.Axis, place.Rotation.Angle * 180 /  math.pi )
-                    else:
-                        nshape.translate(Base.Vector(pts.X,pts.Y,pts.Z))
-                    i += 1
-                    base.append(nshape)
-        obj.Count = i
-        if i > 0:
-            obj.Shape = Part.makeCompound(base)
-        else:
-            FreeCAD.Console.PrintError(translate("draft","No point found\n"))
-            obj.Shape = obj.Base.Shape.copy()
 
 
 ## @}
